@@ -1,7 +1,7 @@
 {
   "name": "@firestitch/form",
   "title": "Form",
-  "version": "13.3.7",
+  "version": "14.0.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/Firestitch/ngx-form"
@@ -28,27 +28,27 @@
     "npm-update": "npx npm-check-updates -u \"/@firestitch/\" --target semver"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular-devkit/core": "^13.3.11",
-    "@angular/animations": "^13.4.0",
-    "@angular/cdk": "^13.3.9",
-    "@angular/cli": "^13.3.11",
-    "@angular/common": "^13.4.0",
-    "@angular/compiler": "^13.4.0",
-    "@angular/compiler-cli": "^13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^13.4.0",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
-    "@angular/router": "^13.4.0",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular-devkit/core": "^14.2.13",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/cli": "^14.2.13",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@firestitch/api": "^12.4.0",
     "@firestitch/autocomplete": "^12.5.3",
     "@firestitch/autocomplete-chips": "^10.2.5",
     "@firestitch/checkboxgroup": "^12.0.1",
     "@firestitch/clear": "^12.4.0",
     "@firestitch/common": "^12.7.0",
-    "@firestitch/component-tools": "^13.0.6",
+    "@firestitch/component-tools": "^14.0.0",
     "@firestitch/core": "^12.1.0",
     "@firestitch/country": "^12.1.0",
     "@firestitch/date": "^12.4.5",
